+++ CHANGELOG.md
@@ -19,13 +19,15 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
   - `getPaywallComponents` now returns a `Result<PaywallComponents>`
 - Removes Gson dependency
 - Adds `isScrollEnabled` flag to enable remote controll of Paywall scrollability
+- Adds `PaywallResourceLoadFail` event to enable tracking of failed resources in Paywall
+- Improves bottom navigation bar color handling
 
 ## Fixes
 - Fixes issue where paywalls without fallback would fail to load and missing resource would cause a failure event
 - Fixes issue with `trialPeriodDays` rounding to the higher value instead of lower, i.e. where `P4W2D` would return 28 days instead of 30, it now returns 30.
 - Fixes issue with system navigation bar not respecting paywall color
 - Fixes issues with cursor allocation in Room transaction
-
+- Improves handling of chromium render process crash
 
 ## 1.4.0-beta.3